CUDA graphs overrides dynamic shapes and forces specialization
Previously, cudagraphs and dynamic_shapes were incompatible and enabling
dynamic shapes would forcibly disable cudagraphs.  This new strategy
I think is better.  The idea is essentially that cudagraphs is an
"optimization" that happens to guard on every input.  When cudagraphs
is on, we force everything static, and this automatically does the right
thing because we will force a recompile if sizes change.

This obsoletes https://github.com/pytorch/pytorch/pull/101813

Signed-off-by: Edward Z. Yang <ezyangmeta.com>

ghstack-source-id: c2d3a6ef2bf7fc4a6fc29df8eeba6174c28e9272
Pull Request resolved: https://github.com/pytorch/pytorch/pull/103290